SNOW-164866 Update JDBC version used by the connector.
Some code changes were required because currently in the connector we close the connection and statement (which in turn closes the result set) before parsing the result set. With the new driver version, this causes an exception to be thrown.

With this change, we do all the processing of the result set while the connection is open.